+++ src/main.py
@@ -1,5 +1,6 @@
 import requests
 from extrator import Extrator
+from termoDocumento import TermoDocumento
 import re
 
 # with open("../links.txt", "r") as arquivo:
@@ -7,4 +8,7 @@ import re
 #         url = url.strip('\n')
 #         if(Extrator().extrair_de(url)):
 #             print("Texto extraido de",url)
-Extrator().contruir_matriz()
+
+tokens = Extrator().extrair_tokens()
+matriz_hash = TermoDocumento().contruir_matriz(tokens)
+matriz_hash_tf = TermoDocumento().preencher_matriz_tf(matriz_hash)
